eth, core: handle state-sync data in eth/69 messages
This PR adds the ability to send state-sync transaction receipts (or bor receipts in short) in eth/69 protocol (which is newly added from upstream). This will allow nodes using snap sync to persist the state-sync transaction receipts (as there's no execution involved in snap sync and hence they're not generated locally).

This PR changes the implementation of 2 p2p messages:
- GetReceiptsMsg (code: 0x0f): Fetches bor receipts for requested blocks and include them in the final encoded message
- ReceiptsMsg (code: 0x10): Accepts the bor receipts from incoming p2p packet. They are excluded from receipt root (in header) and hence this needs to be taken care of while handling incoming receipts.